Refactor: Moved booking types into a new folder
