+++ client/laundroclean/src/components/ui/flexboxes/BookingDisplayTable.tsx
@@ -2,8 +2,8 @@
 import styles from "./BookingDisplayTable.module.css"
 import { FiEdit, FiTrash2 } from "react-icons/fi";
 import { useCompanyUserMenu } from "src/components/layouts/CompanyUser/context/CompanyUserMenuContext";
-import { mapBookingStatus } from "src/types/bookingStatus";
-import { BookingDetail } from "src/types/bookingOrder";
+import { mapBookingStatus } from "src/types/booking/bookingStatus";
+import { BookingDetail } from "src/types/booking/bookingOrder";
 import { mappedDelivery } from "src/services/bookingService/bookingMockData";
 import { transformFieldInArray } from "src/utils/mapData";
