+++ client/laundroclean/src/types/booking/bookingStatus.ts
